backend/requirements.txt
# --- AUTOMATICALLY GENERATED COMBINED REQUIREMENTS FILE ---

# Resolution Strategy:

# 1. Prioritized the higher version for common packages (e.g., fastapi, pydantic, pymongo).

# 2. Included the full stable AI/ML/RAG stack from the second list.

# ===============================================

# 1. WEB FRAMEWORK, ASYNC, AND UTILITIES (Prioritizing newer versions)

# ===============================================

fastapi==0.115.4
uvicorn[standard]==0.32.0  # Used [standard] extra from original list
pydantic==2.9.2
pydantic_core==2.23.4
starlette==0.41.2
typer==0.12.5
click==8.1.7
Jinja2==3.1.4
MarkupSafe==3.0.2
python-multipart==0.0.17
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
sniffio==1.3.1
websockets==13.1
anyio==4.6.2.post1
exceptiongroup==1.2.2
watchfiles==0.24.0

# ===============================================

# 2. DATA, API CLIENTS, AND AUTH

# ===============================================

requests==2.32.3
urllib3==2.2.3
idna==3.10
charset-normalizer==3.4.0
certifi==2024.8.30
PyYAML==6.0.2
python-dotenv==1.0.1
pillow==11.0.0
pymongo==4.10.1

# ===============================================

# 3. GOOGLE AI, RAG, AND ML STACK

# ===============================================

# Google Generative AI & Client Libraries (Using the latest pins/git from the first list)

google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai @ git+https://github.com/google-gemini/generative-ai-python@c72aa6b3c417ad4f91c862de1482b0847b190e6e
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
proto-plus==1.25.0
protobuf==5.28.3

# Core RAG/Vector/ML Dependencies (Pinned to known stable versions from second list)

numpy==1.26.4
torch==2.3.0
faiss-cpu==1.11.0
transformers==4.35.2
sentence-transformers==2.2.2
huggingface-hub==0.17.3

# LangChain and related tools (Unpinned where versions were not specified)

groq
streamlit
streamlit_chat
langchain
langchain-community
pypdf
boto3

# ===============================================

# 4. OTHER UTILITIES AND DEVELOPMENT TOOLS

# ===============================================

annotated-types==0.7.0
cachetools==5.5.0
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi-cli==0.0.5
filelock==3.16.1
httplib2==0.22.0
httptools==0.6.4
identify==2.6.1
lxml==5.3.0
markdown-it-py==3.0.0
mdurl==0.1.2
nodeenv==1.9.1
oauthlib==3.2.2
platformdirs==4.3.6
pre_commit==4.0.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
requests-oauthlib==1.3.1
rich==13.9.4
rsa==4.9
shellingham==1.5.4
soupsieve==2.6
tqdm==4.66.6
tweepy==4.14.0
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.27.1